test: move postgres length regression to functional suite
