+++ test/functional/schema-builder/column/change/change-column/change-column.test.ts
@@ -1,6 +1,7 @@
 import { expect } from "chai"
 import "reflect-metadata"
 import type { DataSource } from "../../../../../../src"
+import { Table } from "../../../../../../src"
 import {
     closeTestingConnections,
     createTestingConnections,
@@ -85,6 +86,69 @@ describe("schema builder > change column", () => {
                 // revert changes
                 nameColumn.length = "255"
                 textColumn.length = "255"
+            }),
+        ))
+
+    it("should alter postgres column length without dropping data", () =>
+        Promise.all(
+            dataSources.map(async (dataSource) => {
+                if (dataSource.driver.options.type !== "postgres") return
+
+                const queryRunner = dataSource.createQueryRunner()
+
+                try {
+                    await queryRunner.createTable(
+                        new Table({
+                            name: "bug",
+                            columns: [
+                                {
+                                    name: "id",
+                                    type: "integer",
+                                    isPrimary: true,
+                                },
+                                {
+                                    name: "example",
+                                    type: "character varying",
+                                    length: "50",
+                                },
+                            ],
+                        }),
+                    )
+
+                    await queryRunner.query(
+                        `INSERT INTO "bug"("id", "example") VALUES (1, 'preserve-me')`,
+                    )
+
+                    const table = await queryRunner.getTable("bug")
+                    const oldColumn = table!.findColumnByName("example")!
+                    const newColumn = oldColumn.clone()
+                    newColumn.length = "51"
+
+                    queryRunner.enableSqlMemory()
+                    await queryRunner.changeColumn(table!, oldColumn, newColumn)
+
+                    const sqlInMemory = queryRunner.getMemorySql()
+                    expect(
+                        sqlInMemory.upQueries.map(({ query }) => query),
+                    ).to.eql([
+                        `ALTER TABLE "bug" ALTER COLUMN "example" TYPE character varying(51)`,
+                    ])
+                    expect(
+                        sqlInMemory.downQueries.map(({ query }) => query),
+                    ).to.eql([
+                        `ALTER TABLE "bug" ALTER COLUMN "example" TYPE character varying(50)`,
+                    ])
+
+                    await queryRunner.executeMemoryUpSql()
+
+                    const rows = await queryRunner.query(
+                        `SELECT "example" FROM "bug" WHERE "id" = 1`,
+                    )
+                    expect(rows[0].example).to.equal("preserve-me")
+                } finally {
+                    queryRunner.disableSqlMemory()
+                    await queryRunner.release()
+                }
             }),
         ))
